Add and apply `black` and `isort` autoformatting
Also add flake8-bugbear.

Because the linting change is applied, there is a lot of whitespace
changing.

black will only install on py3.6+, so it is limited to those platforms,
and flake8-bugbear is similarly limited to py3.5+

Add `make lint` and `make autoformat` as targets to make this work
nicely.

Combine flake8 config with isort config in `setup.cfg`.

After some hesitation, I used platform-specific dependencies in setup.py
I checked and support for these specifiers in setuptools goes back quite
far. Besides, `requests` uses them, so the SDK already needs support.

Jason has been using black and isort for a bit and seems happy with the
results. I'm happy to provide `make autoformat` to autoclean submissions
before people send them to us, and then we can `make lint` and check the
results.

A little bit of cleverness in the Makefile: `black` is only invoked if
it's installed.