Cleanup isort config: apply to __init__ files
Add `__init__.py` files to isort (this is not the default config, but it
will be fixed in a future version of isort). Apply the results.

Also some slight cleanup to Makefile and appveyor config.
I don't see the point in running flake8 linting on appveyor if we aren't
going to just run everything there...