Test use of lowest supported dependencies
Also, revealed that we *actually* require `requests>=2.9.2` because that
version fixed a missing export of an error. This was *not* in the
`requests` changelog, so I'm not super happy about it.

However, this testing ensures that stuff like this can't happen again:
if you start using something which `requests`, `six`, or `pyjwt` adds,
the tests will smack you back down for having the audacity to declare
false compatibility.