Add note to successful transfers docstring
This note is pulled from the underlying API docs. Including it in the
SDK documentation may help avoid confusion about this API behavior in
the future. (I have a case of someone who I believe was tripped up by
the fact that files don't appear in the listing after a sync
transfer.)