Add a sphinx extension for generating method lists
We have a list of methods (like a lightweight index) in many of our
classes. This is fine for smaller classes, but for larger classes
(like TransferClient), it's hard to be sure that the list is complete.

This is easily (enough) solved with a custom sphinx extension which
generates a list of method names which were not inherited or private.
We can then replace all usages of method lists with our custom
directive and be guaranteed up-to-date, alphabetically ordered lists
of methods.