Fix and reorganize docs for material theme
Due to the way that the material theme sidebar displays, it's
desirable to have the tutorial broken down into fewer high-level
sections, and to rename the client section a little to be clearer and
easier to read.

Also fix the logo, remove stale assets from the old docs, and cleanup
some of the text in minor ways.

The only pages which have actually *moved* are the OAuth2 flow pages,
which are now nested under the Auth docs. This keeps the top-level doc
namespace clearer.