Make test matrix smaller, drop 3.5 and add 3.9
Remove "middle python version" testing on Windows and macOS.
The lowest and highest supported versions will be checked on all
platforms, but "other pythons" will only be checked on linux.

This should reduce the amount of CPU we hog, reduce the chances of a
flaky failure, and make testing faster.

Also update classifiers.