Add a pyinstaller hook and simple test
This adds a pyinstaller hook and a simple test to the testsuite which
tries to do a pyinstaller build of a miniscule (two line) app.

The pyinstaller hook is quite simple -- it just ensures that
non-python files are included in the SDK package. The hook itself
needs a setuptools entry point which identifies the hook location (one
of the only bits of setuptools config pyinstaller respects).

In order for the testing to work, a few steps are needed:
- add pyinstaller to 'development' requirements for py3+
- add the test to the testsuite
- skip the test if the environment doesn't support a pyinstaller build

Skipping in the right contexts is the trickiest part. It seems that
pyinstaller needs a shared-library python build. That's the default
for a lot of system packaging, but it's not guaranteed. Furthermore,
pyenv builds won't include shared libraries by default, but will use
static linking instead. Precisely because this is so finicky, we want
to allow tests without it. That way, developer laptop testing can work
even if the python env isn't "just so".

GitHub Actions should build pythons appropriately to run the
pyinstaller test, so we have a good guarantee of things working.

We don't test the pyinstaller hook for py2 -- if it doesn't work for
that version, that's fine (wontfix).