Update pre-commit and dev dependencies
These are pretty innocuous changes to some of the dev dependencies.
Primarily, I want to get the workflow checker added to pre-commit.

I've also opened up `twine` to `<4.0` in the dev dependencies. WAY
back in time, we had issues in which minor twine versions would break
for some of us, so we were pinning the exact version. We can go back
to pinning it if it becomes an issue, but I doubt it will be necessary
again.

Update pre-commit in tox.ini because we now need pre-commit>=2.9