Remove `raw_input` from docs and blacken-docs
This is a basic cleanup of any/all code examples in docs, which I
started on when hunting down py2-compatible content in the docs.

blacken-docs does "the right thing" for us. We need to mark console
blocks as non-python though, or it will fail to parse them.

Also cleanup pre-commit config a bit.