Get mypy linting to pass
1. Add `__init__.py` files to all of the test directories so that mypy
   will treat them as packages

pytest doesn't care, but mypy flips out if two distinct modules appear
to have the same name. Making `tests/` look like a package pacifies
it.

2. Annotate some attributes of BaseClient

BaseClient.error_class should be a Type[GlobusAPIError] for things to
behave correctly. However, without an annotation to this effect, mypy
sees two different classes being assigned to the same variable when
it's reassigned in subclasses. Unable to safely infer the type of
`error_class`, it errors.

Since no runtime checks are introduced, you could still safely
subclass BaseClient and replace `error_class` with a type that doesn't
inherit from GlobusAPIError and it would work.

The same applies to the `default_response_class` and
`allowed_authorizer_types` annotations.

We won't be exposing these annotations to users until/unless we mark
the SDK as a PEP561 [1] package. So we can safely add these
annotations and we don't need to worry about breaking some downstream
consumer's `mypy` runs.

3. Mark some test params with `type: ignore`

4. Cleanup the pyinstaller tests

sysconfig conditional imports are for py2-compatibility, and having a
variable of type Optional[module] was making mypy sad.

[1] https://www.python.org/dev/peps/pep-0561/