Import and adapt globus-sdk-tokenstorage
For the most part, this is

    globus_sdk_tokenstorage -> globus_sdk.tokenstorage

But there are some notable caveats. One is that the StorageAdapter and
FileAdapter are public parts of the module and explicitly set their
metaclass to "abc.ABCMeta" (which marks them as abstract). With that
change in hand we can drop "Abstract" from their names for an overall
tidier appearance.

A StorageAdapter is literally any object implementing `on_refresh`
and `store` methods. And the default implementation of `on_refresh`
just calls `store`. FileAdapter just extends this with some notion
that the storage is in a file somewhere.

The other main thing that is new here is that fairly comprehensive type
annotations are now in place, at least enough to note the argument and
return types of the new interfaces.
I've tried to strike a balance on the annotations between flexibility
(e.g. mapping vs dict) and simplicity. So maybe some methods would
accept types other than `str`, but they're all noted as `str` and
that's what should be used.

The tests are pretty much a verbatim port of the
globus-sdk-tokenstorage tests.

One potential issue is that FileAdapter still only sets umask
appropriately for a unix environment. I would be happy to work on
changes to add windows compatibility, but for now there is no attempt
at doing so. (globus-sdk-tokenstorage does not solve this problem
either)

Also notable: SQLiteAdapter inherits from FileAdapter (this was not
the case in the standalone lib) so that we can control the mode of the
created db via `FileAdapter.user_only_umask()`.