Separate transport (requests) from BaseClient
This introduces a new subpackage, `globus_sdk.transport`, which is
responsible for most of the details of HTTP requests. Ideally, we
could move all logic regarding requests down into it, insulating the
BaseClient from details about how requests are encoded, prepared, and
sent.

The JSON vs Form data dispatch is now handled by this layer (we pass
through an `encoding: str` parameter to control it), as is some simple
retry handling.

The Transport layer is naive and doesn't know about the retries which
may be caused by a 401 -> token refresh. However, this simply means
that the Transport auto-retry behavior applies to the initial request,
the refresh request, and the retried request.

Because the retry behavior includes internal `sleep()` calls to do
backoff or respect a RetryAfter header on a 429 or 503, it's possible
for calls to become very slow in some scenarios.