Refactor retry handling to use chainable objects
Rather than having singular retry checking logic, move this logic into
a suite of "checker" objects, which can be asked in series if the
request should be retried.

Because the checkers are not exclusive, it's very easy for someone to
come along and add a custom checker to the existing suite in the
future. For example, if an API knows that 409s with `code="TryAgain"`
can be retried, that logic can be encoded in a new checker.