Convert 401 handling into a retry checker
This requires some extensions to the RetryContext to pass along a dict
of "state" information, so that a past retry can send "forward" the
fact that it already tried to do a credential refresh.
This avoids multiple (cyclic) refreshes when a service is giving
persistent 401 responses.

Because there are no cases in which retry_401=False is set, this
parameter is removed from the various helpers (it would need to be
passed down to the retry checker otherwise).