Replace internal cfg file with new conf system
Remove the `globus_sdk/globus.cfg` file, whose primary purpose was to
store service URLs for use by the SDK when various environments are
set.

Instead, environments are represented with a new set of EnvConfig
classes, with a registry dict that maps envname -> class.
So we can lookup the `preview` EnvConfig, and then ask it for the
`auth` URL.

None of the rest of the configuration handling has been touched by
these changes -- the config loading from `$HOME` and various options
for verify_ssl, http_timeout, etc. are unchanged.