Strip down the contrib doc and add badges
A lot of advisory guidance in the contrib doc is not essential and
takes up space/time/attention. e.g. It is easiest to handle PRs which
merge cleanly, and commit messages ought to use the imperative mood.
However, these issues are not dealbreakers -- we can afford to rebase
PRs or put up with awkwardly phrased commit messages.

Rather than populating the contrib doc with best practices and
advisory info, cut everything down to just the necessary tox commands.

To get info about code style out into the world, setup badges for
black and isort.