Convert RetryChecker objects to hooks
Rather than having classes and objects in a heirarchy, just make them
callbacks which are kept in a list and invoked in series.
The core logic remains the same, but registering a new hook (or
defining a new policy which removes several or all default hooks)
becomes much simpler.

All of the default hooks are public methods of the base RetryPolicy
and are named `default_check_*`