Update to pyjwt 2.x and fix jwt tests
Add passthrough parameters for "the jwt library" (pyjwt) to
decode_id_token. This allows the testsuite to pass `verify_exp=False`
as one of the pyjwt options.

After testing, this works on both pyjwt v1.x and v2.x , as the option
name is unchanged. Since the testing is done, do the library update as
well.