Consolidate utility methods into utils module
Having a utils/ subpackage is unnecessary -- there are multiple
modules here each holding single functions.
Additionally, there are methods in `base.py` which ought to be in
`utils`. Move everything into this module and then convert the import
style to be `from globus_sdk import utils` internally, so that the
imports are simpler and we have easy access to all utilities.