Type annotate methods on UUIDs in TransferClient
Endpoint ID, Task ID, and Bookmark ID are all UUIDs. They are
therefore annotated with a type of `Union[bytes, str, UUID]` and are
always passed through the safe_stringify method to ensure that we have
a `str` type.

Because this introduces annotations to a large number of methods which
previously didn't have any, include the following adjustments:
- annotate with the return types
- use `typing.cast` to harmonize the return value with the declared
  type
- use a fresh variable name (e.g. `endpoint_id_s`) for the stringified
  version of an argument