Switch to a src/ layout and add py.typed
This layout has benefits when running tests and type-checking. Because
the current directory is always in the import path for python and it's
typical for tox and pytest runs to take place from the repo root,
`import globus_sdk` can easily refer to the local copy, *not* an
installed version. This means that in the presence of a packaging
issue or any bug which behaves differently under a package install (vs
local imports), the local import behavior could mask the problem.

Also, add a dedicated `mypy` tox env which does a `mypy` run against
the whole `src/` directory. This helps avoid issues which can be left
uncaught by `pre-commit` when multiple modules conflict, but
each file is internally consistent.

Add a `py.typed` file, indicating that `globus_sdk` provides type
annotations.