Fix type annotations for client HTTP methods
Introduce a type annotation for the default response type of a
client class, so that `get()`, etc. can be annotated to return that
type. This makes BaseClient a generic which can be parametrized over
the valid return types. In order to notate this accurately with a
default of `GlobusHTTPResponse`, an overloaded `__new__` with
annotations is needed, but the new constructor has no meaningful
content.

Overloads for the various HTTP methods can then be written to produce
`DefaultResponse_T` as the return type when `response_class=None` and
`Response_T` when `response_class=Type[Response_T]`. This gives the
client methods the correct type signatures and lets mypy deduce
behaviors correctly on get/put/post/...
Unfortunately, the annotations need to be repeated for each method.

The most exciting and interesting part of this change is that
`default_response_class` has been converted into a property which
inspects the current class to see if it was a parametrized generic.
That is, when we declare

    class TransferClient(BaseClient[TransferResponse]): ...

we automatically get `default_response_class=TransferResponse` without
having to harmonize a class attribute with the class declaration.

In order for this to work, we need to dig through the type information
on the current class. On py3.6 and py3.7, the behavior is significantly
different from py3.8+, so caution and multipython testing is
essential.