Remove BaseClient.default_response_class
Instead, always return GlobusHTTPResponse unless a `response_class` is
specified.
For SearchClient and AuthClient, this is no change. For
TransferClient, we are left with a choice between updating all HTTP
method calls to specify `response_class=TransferResponse` or removing
the usage of TransferResponse.

It turns out that without these, there's no need for TransferResponse
at all. Remove it, and replace it with GlobusHTTPResponse as the base
class for customized response types.

This simplifies type annotations for BaseClient because a concrete
type replaces a class variable. It no longer needs to be declared as a
generic over response types.