Allow responses to wrap other responses
This rewrite net-simplifies the combination of responses and various
HTTP methods. The overall gist is that base client methods always
return GlobusHTTPResponse objects, and the ability to pass
`response_class` is removed. In order to still support advanced
response types (ActivationRequirementsResponse, OAuthTokenResponse,
etc), GlobusHTTPResponse now supports initialization from some other
"wrapped" response object.

During init, `GlobusHTTPResponse` checks if its input is a
`GlobusHTTPResponse` and dispatches on that for all of its logic.
Either it stores the `_wrapped` response or the `_response` (raw
requests.Response object). The various attributes of the response are
either copied from the wrapped response or are pulled off of the
response object as before.

As a result, where we would previously provide a desirable response
class like so:

    self.get("/foo", response_class=FooResponse)

The same idea is expressed in a more "ordinary" and idiomatic way like
so:

    FooResponse(self.get("/foo"))

The following included changes are related to this:

- PaginatedResponse no longer needs to take a `response_class` to pass
  through -- it just wraps data in IterableTransferResponse

- The once-optional `client` argument to GlobusHTTPResponse is now
  required, since all of this was being refactored anyway

- Type overloads for the BaseClient methods are all removed, because
  they now are guaranteed to always return GlobusHTTPResponse

- AuthClient.oauth2_token needs to allow the use of variable wrapper
  types to support dependent token responses while still returning a
  default of an OAuthTokenResponse. It therefore implements and has
  type overloads for `response_class`, similar to what was previously
  done in BaseClient