Clean up the authorizers subpackage
This is a combination of interface changes and internal cleanup.

Rather than authorizers having to modify a header dict (assumes
headers are a dict, assumes that modifying that dict is the right way
to get authz info), authorizers now just return a valid authz header.
It's up to other code, like the SDK transport, to decide how to
incorporate that info into a request. As such,
`set_authorization_header` has been renamed to
`get_authorization_header`.

Remove unnecessary un-bounded `**kwargs` from the definition of
handle_missing_authorization.

Renewing Authorizers now throw errors instead of logging warning when
initialized with bad data.

RenewingAuthorizer.check_expiration_time is renamed to
RenewingAuthorizer.ensure_valid_token.

RenewingAuthorizer.expires_at is a verbatim value. The expiration
epsilon of 60s is applied when the time is compared, not when the
value is assigned.

Renewing authorizers compute the token hash "automatically" by making
`access_token` a property with a custom setter.

Loggers are renamed to `log` for uniformity. Local relative imports
are used in the subpackage. Format strings and logging are cleaned up
and condensed.