Combine RetryPolicy and RequestsTransport
All of the properties of the retry policy are included in the
transport, and it is solely responsible both for request sending and
handling retries. As a result, more options are available as retry
params via the `transport_params` argument to client construction.
It's still possible to set various settings with a custom
`transport_class` as well.

i.e. Both imperative and declarative styles are supported well by the
reduction in distinct objects.

Update changelog to include this change.