Add `items()` iteration to paginators
Support both `pages()` and `__iter__` usage, consolidate the relevant
logic into the base Paginator class, and make `items()` an iterator
over items in `pages()` when the `items_key` is set for the paginator.

All currently annotated `has_paginator` methods are marked with an
`items_key`. Trying to use `items()` when `items_key` was not set is a
simple ValueError.