Fix paginators with 'max_results_per_call' (#431)
This argument is not valid -- these appear to be typos where the
desired parameter was `max_total_results`. After validating with the
API, this appears to be correct: these APIs for task_list and
task_event_list support a max page size of 1000 and a total limit of
1000 (meaning you can get all results in a single page).

Add testing for `task_list` to ensure that the relevant paginator is
constructed and iterated.