Add parameters for ls, autoactivate, ep search
Several parameters were not explicitly supported in SDK v1.x and 2.x
because we could rely on `**params` passthrough. Now that `**params`
usage is gone, it's more important to have these methods explicitly
supported as parameters.

In addition to adding the params, include some basic tests which
ensure that they are parsed and added to query params correctly. (Not
done exhaustively, just on the basics).