Fix tests to always set encoding on mock responses
Newer versions of encoding detection warn about small sample sizes
when guessing encoding (`charset_normalizer` after 2021-07). To avoid
these warnings, just make sure that the encoding is always set. Use
utf-8 and ISO-8559-1 in tests.

UTF-8 is the most commonly set encoding via 'charset' in a
content-type header, and ISO-8559-1 is the default encoding specified
for an HTTP body when there is no explicit encoding given in a charset
header (though, of course, many servers are not spec-compliant).