Convert data helper classes to "PayloadWrapper" (#438)
* Convert data helper classes to "PayloadWrapper"

Rather than defining classes which inherit from `dict` directly,
define PayloadWrapper, a utility class which wraps a `_payload` dict.
The PayloadWrapper inherits from MutableMapping and implements methods
against its underlying dict, meaning that the range of usages for
these objects is not diminished.

TransferData, DeleteData, and SearchQuery are all updated to inherit
from this object and, other than invoking super().__init__(), they are
almost entirely unchanged.
In the case of SearchData, __init__ needed to be defined to allow
passing known parameters as keyword args. (Some previously valid
usages will now require the use of `additional_fields`).

In terms of methods which pass data as a dict, the changes made are to
allow any PayloadWrapper object to pass down to
BaseClient.request(...). At that stage, `PayloadWrapper`s will render
with a `to_dict()` call. As a result, almost no concrete code, only
type annotations. Mostly changes of the form
    `Optional[Dict]` -> `Union[None, Dict, PayloadWrapper]`

A new test exercises the methods of PayloadWrapper to make sure they
all function as expected, and a minor update ensures that SearchQuery
is well-tested.

The Transfer data module has been split in two, to make it easier to
read/maintain the two helper classes.

* Use UserDict to implement PayloadWrapper

Swapping MutableMapping for UserDict lets us remove all method
definitions, which is nice. Keep the PayloadWrapper unit test, but
cleanup the comments.

A comment on PayloadWrapper explains why we should prefer it over
subclassing dict directly.