Refactor error parsing to have clearer flow and handoff to "ErrorInfo" objects (#441)
* Enhance and improve error parsing

GlobusAPIError gains a new property, `info`, which contains variables
for `authorization_parameters` and `consent_required`, each of which
points to an "ErrorInfo" object of the appropriate type.

This encapsulates any parsing of the response data (beyond basic JSON
parsing) in a dedicated set of classes.

Because the `.info` attribute is added to all errors, the information
about the type of error (e.g. "consent required") is independent from
the particular API error class.
Effectively, this is composition ("has a") rather than inheritance
("is a").

Additionally, the "base" error parsing logic can be simplified. Start
by assuming that `code` and `message` are, as for text responses,
"Error" and <response>.text . Then overwrite them when parsing JSON
*if* we find relevant properties.
This reduces the risk that parsing fails while handling an API
response with some -- but not all -- of our expected properties. All
fields are populated, and it's possible to pass `details`, `message`,
and `code` independently. (e.g. an error with `message` but no `code`
will be handled gracefully)

* Expand testing for ErrorInfo parsing

Add several tests and adjust test-cases to ensure that ErrorInfo
objects and their contents are properly exercised.

* Ensure ErrorInfo objects are documented

Add ErrorInfo and the surrounding classes to the exceptions doc page.

Rename ErrInfoContainer to ErrorInfoContainer to be more consistent in
the documentation (and in our class names in general).

* Apply suggestions from code review

Co-authored-by: Rudyard Richter <rudyardrichter@users.noreply.github.com>

* Refactor error parsing per review

Generally aim for "flatter" structure in error parsing.

Co-authored-by: Rudyard Richter <rudyardrichter@users.noreply.github.com>