Add decorator for external doc links
The decorator adds an "External Documentation" section to the end of a
method docstring, useful for linking various client methods to the
appropriate API docs.
There is a default format for the section which can be overridden, and
a default base URL of "https://docs.globus.org/api", which also can be
overridden.

In the case of the GroupsClient docs, which would need to consistently
repoint the base URL to the redoc site, a specialized decorator wraps
the standard one.