Support UUIDLike types for GroupsClient
GroupsClient and GroupsManager both have args that are group_id,
identity_id, or (in one case) parent_id, referring to a parent group.
In all these cases, a UUIDLike should be accepted.

Because this now means that the safe_strseq_iter helper could be
called on a bytes object, it needs slight adjustment.