Add first version of GCSClient
GCSClient supports only one operation (collection listing). Test
rigging is put in place with a test for collection listing and one for
error parsing.

The iterable GCS response implementation overlaps with
IterableTransferResponse, so the commonality is factored out into a
shared mixin class for responses.

One change of broader scope: the GCSResponse class in the gcs-cli
currently defines `__str__` of a response object as an indented
JSON-dump of its data attribute. Rather than making this a specialized
behavior, this is a reasonable improvement for all response objects.
str() of a response will JSON dump the data, and repr() remains the
same. (This makes str() and repr() different.)