Get nitpicky, Werror sphinx builds to pass
The sphinx build now runs with `-W`, which sets warning-is-error on
the build. Sphinx builds with warnings will no longer pass.

The tox config has been modified to allow posargs to be passed to the
docs build, so that `tox -e docs -- -n` will set `-n` (nitpick mode).
With nitpick enabled, any cross reference which fails to resolve will
be flagged as an error. This is very aggressive, and leads to many
spurious warnings.

For this reason, `-n` is omitted from the standard build.
However, add a list of nitpick ignore rules which get the build to
pass under `-n` for now. This is not guaranteed to keep working, but
the nitpick build did discover several missed or malformed
annotations. It also flags type annotations which are perfectly
correct/valid, but which sphinx can't use for correct cross references
because of their ambiguity.

The state-of-the-art with respect to nitpick_ignore and
nitpick_ignore_regex appears to be to load these data from a file in
the docs dir. This is the approach taken by astropy [1], and it is
shamelessly imitated here. Regex ignores are so useful, they are
included in this implementation.

[1] astropy doc nitpick ignores:
https://github.com/astropy/astropy/blob/a7cf66db8ed95ca448ffb5a4c65f0a4a681ec220/docs/conf.py#L292-L297