Add GCS Get Collection support
Add a variant of the SingletonGCSResponse, which provides access to
the original JSON data as well as the extracted data attribute.

Add a new method, GCSClient.get_collection(ID), which returns a
SingletonGCSResponse.

Update the usage on get_collection_list to handle `include` as a
`str | Sequence[str]`, and update docstring to cover the parameters.

Testing covers SingletonGCSResponse on expected and "malformed" but
readable data.