Make args to endpoint_manager_task_list explicit
Rather than putting all of the various supported options in
`query_params`, make each filter type a keyword argument, annotated
with supported types.
Each argument is tested and passed into the inner dict as appropriate,
in some cases with some encoding or comma-joining steps.

Tests do not worry about response payloads, but ensure that the
expected parameters are sent as query params.

Docs are converted from a table to standard `:param:` docs. Some
wording was adjusted, but mostly it is copy-paste and reindent.