Replace Singleton with UnpackingGCSResponse
SingletonGCSResponse looked to unpack data from a list of length
exactly one. After asking the GCS team about this, the real need is
for a way of unpacking an array which may include additional data via
the `include=...` query params. For this reason, the better approach
is to have an "UnpackingGCSResponse" type which traverses the `data`
array and pulls out the first matching element.

The norm is matching on the DATA_TYPE field, but we need to accept
`collection#1.0.0` and `collection#1.1.0` because the document types
are semver versioned. To support this, we can take a string regex for
matching, and then `collection#1\.\d+\.\d+` is accurate for matching
any 1.x version.

We may want to refine this further in the future, and we may have more
complex matching requirements. For this reason, as an alternate to the
string -> regex -> DATA_TYPE matching approach, UnpackingGCSResponse
can accept a callback which is used on elements of the `data` array to
test for whether or not a piece of data should be "unpacked". The
logic is still "first match wins", so anything more sophisticated than
that would need to be done in a completely new custom response type.