Add method to get local GCP owner from gridmap
Minimally parse the gridmap file, find a GCP DN, and then handle the
result. It can be a globusid.org username or auth ID, so these are
flagged on return. If users want a uniform(ish) interface, they can
pass in an AuthClient, which is used to do an get-identities call to
lookup the user info.

Skip coverage reporting on overloads for more accurate coverage info.