Fix PermissionError test for windows
shutil.rmtree on windows doesn't handle 0-permissioned files
automatically, so it blows up the testsuite. However, the tests also
fail because chmod doesn't "do the right thing" on windows, as there's
no unreadable file attribute which can be set.

To work around this, make the 'chmod 000' behavior non-windows only,
and use a directory (rather than a file) for windows testing. Because
directories have attributes that aren't compatible with opening for
reading, this has the desired effect for the purposes of testing.