Add filter_is_error to advanced task event list
The param is added as a bool. However, testing covers setting it as an
int. This is the usage which was seen in the wild under SDK v2.x which
prompted this addition, and checking that it is supported (although
not documented and not to be encouraged) is simple enough.