Create GCP OwnerInfo objects
Instead of `(str, bool)` tuples, create a dedicated object for this
purpose. Documentation covers its usage and members, custom `__str__`
shows it nicely, and tests cover the behavior.
A classmethod of the info object constructs it from a given filename
with very limited error handling, and the
LocalGlobusConnectPersonal.get_owner_info() code handles the relevant
errors and converts them to None.

Due to its increased size, the owner info class is split into a
separate module which covers that class as well as the B32 parsing
utilities.