Add support for Collection POST,PATCH
These are the create and update operations for GCS Collections.
There is also API support for PUT, but as it is harder to use
correctly than PATCH, and does not support operations which cannot be
accomplished via the PATCH endpoint, no SDK method is added for it
here. One can easily be added later if necessary.

With respect to the payload helpers, this introduces three new
classes: a base class, and a class for each of Mapped and Guest
collection types.

Because the same payload helpers will be used for POST and PATCH
behaviors, none of the arguments are required. Documentation covers
the fact that some arguments are needed during creation.

Many arguments are repeated verbatim between the mapped and guest
collection types and their base class. For this reason, the `dead`
code analyzer is particularly helpful. Unlike code coverage analysis,
or `pycqa`'s unused variable checks, `dead` will insist that each
variable in a function definition is used. However, `dead` is easily
misled, e.g. by signature lines which are needed to match some
external specification. For this reason, `dead` is being enabled as a
selective linting step, only against the GCS helper module for now.

Some unusual conventions around commenting on options and specifying
an ordering are used to ensure that the module is internally
consistent, and they are documented in a block comment in that module.

GCS API requires that the `DATA_TYPE` field be set based on the newest
field used in a payload. That is, if `foo` was added in v1.5.0, then
the `DATA_TYPE` field must be v1.5.0 or higher. However, SDK users and
clients (e.g. globus-cli) will want to maintain compatibility with as
many GCS servers as is possible at any given time. For this reason,
`DATA_TYPE` should not be aggressively set to the highest known
version. In support of these requirements, `CollectionDocument` will
inspect the fields which are present and set `DATA_TYPE` to the lowest
possible version number. In support of new fields which require higher
version numbers, this is exposed as a modifiable mapping on the
CollectionDocument, `DATATYPE_VERSION_IMPLICATIONS`. Additionally,
`DATA_TYPE` can be passed explicitly to set a known value. This usage
is to be considered secondary to use of
`DATATYPE_VERSION_IMPLICATIONS`, which is the preferred way to set
`DATA_TYPE`.
See `test_datatype_version_deduction` and related tests for a
demonstration of this behavior.