Replace regex match in UnpackingGCSResponse
Instead of passing `1\.\d+\.\d+` to match a version range, this
improves the interface to allow passing a simple version spec.

`collection>=1,<2` will check for `collection#x.y.z` where `x=1`, and
so forth. The version specs work basically the same way that pypa
package version specifiers work, and version specifiers for most
packaging systems.

In order to parse these, a module is added which splits up the string
and converts it into a "version range". Tests cover all of the major
conditions.