Add support for specifying config_dir to local GCP
The local endpoint object now supports `config_dir: str`. Because this
value is detected when `None`, it effectively replaces some of the
`_local_data_dir` logic with something a bit more readable.

In order to align with GCP Linux's `-dir` option, the directory on
non-Windows platforms is `~/.globusonline/`, not
`~/.globusonline/lta`. This means that there is still a local data dir
property which joins with `"lta"` on non-Windows platforms.

(Don't blame the SDK for this. It's a very, very old GCP behavior and
we're just aligning with it here.)