Remove 'safe_stringify'; remove support for bytes
Removing support for `bytes` as part of a UUIDLike generally makes our
internal code simpler. Because `UUID` and `str` can both be used if
passed through `str()` and string formatting defaults to using `str()`
on elements, in many contexts no conversion is needed now. (A UUID can
be used in an f-string and gets formatted correctly.)

There's also some minor cleanup in DeleteData to be more like
TransferData.