Fix type-ignore on AuthClient by setting bound
When the type var we use for `response_class` is unbounded, `mypy`
gives an error because `object` does not support construction from a
response. However, the correct bound for this type is
`GlobusHTTPResponse`, which ensures that only types which can be
constructed from a response may be used for `response_class` in
`oauth2_token`.