Use casts in decorators instead of `type: ignore`
Casting to a stub class, rather than using `type: ignore` comments, is
exactly the same in meaning.

However, it means that we no longer have any type-ignores in the
`src/` tree (good) and it statically declares the attributes somewhere
outside of the decorators. If we wanted, we could cast in other
contexts for access.