Add testing for package metadata
1. Add a `twine-check` tox env
2. Add a `poetry-check` tox env
3. Run both of these under CI

twine-check runs the `twine check` command after a package build.
poetry-check requires that we define some non-pytest testing data, so
an appropriate dir is setup with the test files inside of it. The
`poetry-lock-test` contains a pyproject.toml where `poetry lock`
should work and point at the current version of the SDK in-tree. If
the package metadata is not correct (as it was in #474), the
`poetry lock` command will fail.