Update the minimum requests and crypto versions
This sets a lower bound on `cryptography` (missing), and adds a
`py36-mindeps` env to tox. Run in this mode, we verify the SDK
testsuite against the minimum dependency versions we specify, to
ensure compatibility.

Also, import RSAPublicKey from the module where it is declared, in the
hopes that this will be more stable.

resolves #481