Remove the job which commits changelog fixes
This is currently not going to work because you can't push to a fork
branch using GITHUB_TOKEN.

One way to get it working again is to figure out how to map a new file
to a PR number (hard? easy?) and then run the fixer on the changelog
fragment. If this is viable, the existing work here might be
salvageable.