Fix type-preservation for internal decorators
Several utility decorators in the SDK were preserving return types,
but none of the rest of the type signature for the functions and
methods which they decorate.

To resolve, specify them in a more appropriate way that indicates to
type-checkers that the type of the original function is being
preserved.