Adjust cryptography bounds to handle pyjwt
pyjwt>=2.0 requires cryptography>=3.3.1
Rather than specifying a lower bound than that, raise our minimum
version to match. This signals to us, to our users, and to dependency
solvers that we need this version.

On `cryptography<3.1`, `decode_id_token()` fails because PyJWT tries
to load an RSA pubkey and uses the `cryptography` API in a way that
was newly added in v3.1 (does not pass a backend).