Add resource_server as a BaseClient class property
The basic idea is very simple: `BaseClient.resource_server` should
be a read-only proxy attribute which reads from
`BaseClient.scopes.resource_server`.

Unfortunately, we won't have a working version of classproperty from
the stdlib until python 3.9, so we must define our own.
This is not terribly hard once you've done the reading, but it is
extremely confusing until then.

A descriptor is an object which implements `__get__`, `__set__` and
`__delete__`. Only `__get__` is strictly required, and there are odd
precedence rules which are driven by the presence of `__set__` or
`__delete__`. The python docs recommend defining `__set__` to raise an
AttributeError in order to define a read-only descriptor.

Both `classmethod` and `property` are descriptors, and the python docs
happily provide examples of pure python versions of these descriptors.
Combining the two into a read-only classproperty descriptor follows
from there.

Naming-wise, I went with `sdk_classproperty` in the utils module. It's
short enough, and is unlikely to conflict or be confused with some
other `classproperty` descriptor which anyone else might use.

The last significant hurdle here is type annotation. The descriptor is
an object which binds a method in __init__ and calls that method in
__get__, meaning that it has to be a Generic in order to preserve a
TypeVar across those two definitions. But wait, there's more! Since we
took in an ordinary looking method and turned it into a classmethod
(ish), we need to explain to mypy that the original getter may have
operated on `T`, but the reformulated one operates on `Type[T]`. So we
have a Generic parametrized over two types: the class where the
descriptor is applied, and the return type for the getter.

Finally, all done, add tests for the few cases of this which matter.
Primarily, we want to confirm that `GCSClient.resource_server is None`
and that `TransferClient.resource_server` is
`transfer.api.globus.org`.

In the descirptor, I've added an additional strong warning that
nobody external to the SDK should use that descriptor. That's already
the nature of `utils`, but we've also seen people using internal
methods frequently. Hopefully the extra admonition will help avoid
any mishaps.